requirements.txt
# Mimic - Browser Automation with Muscle Memory
# Install with: uv pip install -r requirements.txt

# Core browser automation
browser-use>=0.2.6
cdp-use>=1.4.4

# LLM providers
openai>=2.8.1
anthropic>=0.75.0
google-genai>=1.53.0
groq>=0.37.0
ollama>=0.6.1

# LangChain (for LLM provider abstraction)
langchain-openai
langchain-anthropic
langchain-google-genai

# Environment and configuration
python-dotenv>=1.2.1
pydantic>=2.12.5
pydantic-settings>=2.12.0

# Async utilities
aiohttp>=3.12.15
aiofiles>=25.1.0
httpx>=0.28.1

# Document handling
beautifulsoup4>=4.14.3
lxml>=6.0.2
pypdf>=6.4.0
markdownify>=1.2.2

# 2FA / TOTP support
pyotp>=2.9.0

# CLI
click>=8.3.1
rich>=14.2.0
typer-slim>=0.20.0

# Image processing
pillow>=12.0.0

# Telemetry (optional)
posthog>=7.0.1

# Development dependencies (optional)
# pytest>=9.0.1
# pytest-asyncio>=1.3.0
# pytest-httpserver>=1.1.3
# ruff>=0.14.7
# pyright>=1.1.407
# pre-commit>=4.5.0
